AGT_platform/backend/requirements.txt
flask==3.0.3
flask_cors==4.0.1
sqlalchemy==2.0.32

alembic==1.13.2
python-dotenv==1.0.1
authlib==1.3.1
requests==2.32.3
pyjwt==2.9.0

celery==5.4.0
redis==5.0.8

boto3==1.35.36

nbformat==5.10.4
pypdf==5.0.1
psycopg2-binary==2.9.11